added resizing, crosshair, and prototype for how to get optimized walls

+++ game/render/render_calc.c
@@ -1,7 +1,6 @@
 #include "../../include/cub3d.h"
 
-
-double normalize_angle(double angle)
+float normalize_angle(float angle)
 {
 	while (angle < 0)
 		angle += RA;
